Move parseInput to a new function
