+++ server/calcapi.ts
@@ -1,81 +1,22 @@
 import express from 'express';
 import execCalc from './calculator/execCalc';
-import exprMap from './calculator/exprMap';
 import { gaussianElimination, Matrix, RandomMatrix, Vector } from './calculator/models';
 import shuntingYard from './calculator/shuntingYard';
+import parseInput from './parseInput';
 import questions from './questions';
-import { IQuestionOptions, TCalc, TQuestion } from './types';
-
-function joinRegex(...regex: RegExp[]): RegExp {
-    return new RegExp(regex.map(item => item.source).join('|'), 'gi');
-}
+import { IQuestion, IQuestionOptions, TCalc, TQuestion } from './types';
 
 const router = express.Router();
 
 router.use(express.json());
 
 router.get('/', (req, res) => {
-    const NUMBER_REGEX: RegExp = /(-1)|[\d\.]+/;
-    const EXPR_REGEX: RegExp = /[a-z*+/\^]+/i;
-    const MATRIX_REGEX: RegExp = /\[(\[(-?[\d\.]+,?)+\],?)+\]/;
-    // same as matrix but with only 1 inner array
-    const VECTOR_REGEX: RegExp = /\[(\[-?[\d\.]+,?\],?)+\]/;
-    const BRACKET_REGEX: RegExp = /[\(\)]/;
 
     let { query: {calc: body} } = req;
-
     if (!body && typeof(body) !== 'string') return res.status(400);
 
-    let partsArr: string[] | null = body
-        .replace(/\s/g, '')
-        // -n = -1 * n
-        .replace(/-(?![^\[]*\])/g, '+-1*')
-        // Remove a plus if it is at the start of a statement
-        .replace(/(?<=^|\()\+/g, '')
-        .replace(/([\d\]\)])([\(\[])/g, '$1*$2')
-        .match(joinRegex(
-            NUMBER_REGEX,
-            EXPR_REGEX,
-            MATRIX_REGEX,
-            BRACKET_REGEX
-        ));
-
-    console.log(partsArr);
-
-    if (!partsArr) return res.sendStatus(400);
-
     try {
-        let calc: TCalc[] = partsArr.map(item => {
-            // Vector must come first as matrix will also match vectors
-            if (item.match(VECTOR_REGEX)) {
-                return {
-                    type: 'vector',
-                    // Regex means that there will only be 1 inner array
-                    data: new Vector(JSON.parse(item).map((row: number[]) => row[0]))
-                };
-            } else if (item.match(MATRIX_REGEX)) {
-                return {
-                    type: 'matrix',
-                    data: new Matrix(JSON.parse(item))
-                };
-            } else if (item.match(NUMBER_REGEX)) {
-                return {
-                    type: 'no',
-                    data: +item
-                };
-            } else if (item.match(EXPR_REGEX)) {
-                if (!Object.keys(exprMap).includes(item)) throw new Error('invalid expr');
-                return {
-                    type: 'expr',
-                    data: exprMap[item]
-                };
-            } else {
-                return {
-                    type: 'bracket',
-                    data: item
-                };
-            }
-        });
+        let calc: TCalc[] = parseInput(body);
 
         let postFix: TCalc[] = shuntingYard(calc);
         let result: any[] = execCalc(postFix).map(result => {
