+++ server/calculator/exprs/index.ts
@@ -1,6 +1,7 @@
 export { default as GetDet } from './getDet';
 export { default as Eigen } from './eigen';
 export { default as Diagonalize } from './diagonalize';
+export { default as Transpose } from './transpose';
 export { default as Exponent } from './exponent';
 export { default as Addition } from './addition';
 export { default as Multiplication } from './multiplication';
